cart-service/requirements.txt
django
djangorestframework
requests
dj-database-url
gunicorn
